Error codes are always negative.
This is to avoid confusion between the From instance and the error code that is
ultimately displayed to the user.

The remaining issue is that the variant tags are still positive since we cannot
easily modify them in a derive macro. We'd have to have another attribute macro
to achieve this automatically.